Pre-release commit
Rationalized structure by moving everything into one file (probably not
necessary; the reason pyinstaller didn't initially work was that it
could not find the png  files. Solution: make sure that QtDesigner is
working on the correct file (was 'secretly' working on an old blits
one), and make use of the resources option (all pic files must be
resources, and a qrc file must be created).
Then translate the .ui and .qrc files into .py files (using pyuic5 and
pyrcc5) and import the .py equivalents into the driver (biscuit.py)